Refactoring

+++ src/display.rs
@@ -6,11 +6,11 @@ pub trait DisplayInCui {
 
 impl DisplayInCui for LifeGame {
     fn display(&self) {
-        let field_as_string = (0..self.height as i32)
+        let field_as_string = (0..self.height())
             .map(|y| {
                 format!(
                     "{}\n",
-                    (0..self.width as i32)
+                    (0..self.width())
                         .map(|x| match self.get_cell_state(x, y) {
                             CellState::Alive => "■",
                             CellState::Dead => "□",
